requirements.txt
MetaTrader5>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.14.0
scipy>=1.10.0
scikit-learn>=1.3.0
SQLAlchemy>=2.0.0
PyYAML>=6.0
requests>=2.31.0
websockets>=12.0
python-dotenv>=1.0.0
PyQt5>=5.15.0
scikit-image>=0.21.0
ta-lib>=0.4.0
tqdm>=4.65.0
tabulate>=0.9.0
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.0.0
black>=23.0.0

